+++ src/fireball/fireballs.cpp
@@ -15,6 +15,9 @@
  * Code to move, render and otherwise deal with fireballs.
  *
  * $Log$
+ * Revision 1.5  2004/07/04 11:31:43  taylor
+ * amd64 support, compiler warning fixes, don't use software rendering
+ *
  * Revision 1.4  2002/06/17 06:33:08  relnev
  * ryan's struct patch for gcc 2.95
  *
@@ -696,7 +699,7 @@ void fireball_set_framenum(int num)
 	// valid lod?
 	fl = NULL;
 	if((fb->lod >= 0) && (fb->lod < fd->lod_count)){
-		fl = &Fireball_info[Fireballs[num].fireball_info_index].lod[fb->lod];
+		fl = &Fireball_info[Fireballs[num].fireball_info_index].lod[(int)fb->lod];
 	}
 	if(fl == NULL){
 		// argh
